Refactor TablesPanel (NodeBal) to use ExtendedExpansionPanels
