+++ src/components/ExtendedExpansionPanel/index.ts
@@ -1,0 +1,3 @@
+export {
+  default
+} from './ExtendedExpansionPanel';
